+++ src/Utils.js
@@ -21,6 +21,24 @@ const isFunction = (value) => typeof value === 'function';
 const outToConsole = (...args) => process.stdout.write(args.join(' ') + '\n');
 
 /**
+ * Return the path of file relative to a directory.
+ *
+ * @param {string} file
+ * @param {string} dir
+ * @return {string}
+ */
+const pathRelativeByPwd = (file, dir = process.cwd()) => {
+  let relPath = file;
+
+  if (file.startsWith(dir)) {
+    relPath = path.relative(dir, file);
+    if (!path.extname(file)) relPath = path.join(relPath, path.sep);
+  }
+
+  return isWin ? pathToPosix(relPath) : relPath;
+};
+
+/**
  * Parse resource path and raw query from request.
  *
  * @param {string} request
@@ -115,4 +133,5 @@ module.exports = {
   isFunction,
   toCommonJS,
   outToConsole,
+  pathRelativeByPwd,
 };