+++ test/manual/watch-dependencies-bootstrap/webpack.config.js
@@ -1,0 +1,64 @@
+const path = require('path');
+const PugPlugin = require('../../..');
+
+module.exports = {
+  mode: 'production',
+  stats: 'minimal',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+    clean: true,
+  },
+
+  entry: {
+    index: './src/views/home.pug',
+  },
+
+  plugins: [
+    new PugPlugin({
+      js: {
+        filename: 'js/[name].[contenthash:8].js',
+      },
+      css: {
+        filename: 'css/[name].[contenthash:8].css',
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: PugPlugin.loader,
+        options: {
+          method: 'render',
+        },
+      },
+      {
+        test: /\.(css|sass|scss)$/,
+        use: ['css-loader'],
+      },
+      {
+        test: /\.(png|jpe?g|svg|webp|ico)$/i,
+        type: 'asset/resource',
+        generator: {
+          filename: 'assets/img/[name].[hash:8][ext]',
+        },
+      },
+    ],
+  },
+
+  // enable HMR with live reload
+  devServer: {
+    //hot: false,
+    static: {
+      directory: path.join(__dirname, 'dist'),
+    },
+    watchFiles: {
+      paths: ['src/**/*.*'],
+      options: {
+        usePolling: true,
+      },
+    },
+  },
+};